Draft
A new WAV with a full instrumental draft. Still need to fix bass sounds, layer with synths, smooth a couple transitions, and track vocals (70% or so written as of pushing). But this should provide a very good idea of Brighter Days. Hopefully I can get this mostly finished around midday tomorrow and make headway on another track over the weekend!